view : 수동발급 입출력 추가

+++ src/main/java/lotto/model/Money.java
@@ -4,14 +4,14 @@ import java.util.Objects;
 
 public class Money {
 
-    private final int amount;
+    private final long amount;
 
-    public Money(int amount) {
+    public Money(long amount) {
         validate(amount);
         this.amount = amount;
     }
 
-    public int amount() {
+    public long amount() {
         return this.amount;
     }
 
@@ -23,7 +23,7 @@ public class Money {
         return (double) this.amount / other.amount;
     }
 
-    private void validate(int amount) {
+    private void validate(long amount) {
         if (amount < 0) {
             throw new IllegalArgumentException("금액은 음수일 수 없습니다.");
         }
